[US_ND] Make three table YAMLs always historical (Recidiviz/recidiviz-data#32415)
## Description of the change

Sets `always_historical_export: True` in the raw data config files for
`docstars_offendercasestable`, `docstars_offenders`, and
`elite_externalmovements`. We will be adjusting the transfer early next
week, and I don't want to merge this before the files we receive are
actually fully historical because that would cause ingest to treat
incremental files like full historical files, which would cause data to
seemingly disappear downstream (see
https://github.com/Recidiviz/recidiviz-dashboards/issues/5666 for an
example).

These files were modified in the transfer on Monday, so they will arrive
in full historical versions for the first time overnight going into
Tuesday. The plan is to merge this after the first couple of full
historicals have come in.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#32394

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a58de778b18b5829dc378fa8a0b5f265ff6b36aa