[Case Note Search] Remove summary-related fields and turn results field into a JSON for the CaseNoteSearchRecord table. (Recidiviz/recidiviz-data#32604)
## Description of the change

Remove all summary-related fields from the CaseNoteSearchRecord table
and convert the results-related fields to a JSON which stores all result
data. We will nullify the full case note from the response JSON since
that would be a lot of redundant data to store.

Note: In preparation for this non-backwards-compatible migration, I
cleared all rows from the staging/prod CaseNoteSearchRecord databases.

## Testing
Ran this migration locally and tested the write logic on the local db.

GitOrigin-RevId: d8af25c94d491c0565bbdff25c6d1eaf70a16691