[US_TN] Add cleaning to raw data in roster to properly construct supervisor periods due to manual entry differences (Recidiviz/recidiviz-data#32639)
## Description of the change

We receive the raw data to construct StaffSupervisorPeriods from TN as a
manually generated excel sheet which can have various manual entry
errors. Previously, this view did not take into account if someone
entered a SupervisorID as lowercase instead of uppercase letters, which
meant that portions of the ID were cut off in the final regex at the
bottom. Though there are several approaches to fixing this, cleaning the
data as soon as we use it in the view seems to be the cleanest way to
make sure that any other logic we do to match IDs later on will work, so
we added in UPPER to clean the IDs for external_id and SupervisorIDs
when the roster raw data is called in the first 2 CTEs.

This change results in 1556 StaffSupervisorPeriods, down from 1590
previously which is a result of less period changes when the entries
were inconsistent in casing before.

I confirmed with a sandbox that this view successfully runs through the
pipeline and eliminates all references to SupervisorID '01' which was
causing this validation to fail.
https://admin-panel-prod.recidiviz.org/admin/validation_metadata/status/details/current_supervision_staff_missing_email?stateCode=US_TN

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#31430

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: baf9efc5bdbfa8c3a81fe33dbd0d3ee8aaee20cf