[entities] add comment, todo for `convert_schema_object_to_entity` and update usages (Recidiviz/recidiviz-data#32637)
## Description of the change

in Recidiviz/recidiviz-data#32430, i described a few issues i was having with
`convert_schema_object_to_entity` and `populated_backedges`. while that
the sort of backedge conversion is not what
`convert_schema_object_to_entity` was designed for, i realized had been
approaching it's usage quite wrong when anna pointed out (as always, all
shoutouts to anna) that we would prefer to do the sort of tree traversal
i wanted while we still had access to the schema objects (shoutout
sqlalchemy); if we wnat to access something higher up the tree, just
make sure we are converting that object instead of one down the tree.

to that end, in this PR i:
- add a docstring to `convert_schema_object_to_entity` with something
similar to the above
- `populate_back_edges` -> `populate_direct_back_edges`: in the
`convert_schema_object_to_entity` context, we only explore _direct_
backedges and will fail if any object has a backedge that is not
included in the set of the `|schema_object|`, `|schema_object|`'s
forward edges and `|schema_object|`'s direct backedges.
- updated usages of `convert_schema_object_to_entity` w/
`populate_direct_back_edges=False` for calls that would fail due to the
larger backedge network

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#32430

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a246b2b83d4160e47c70b4831ed2def2926d1626