Add full_observability option for cohort methods (Recidiviz/recidiviz-data#32581)
## Description of the change

As noted in Recidiviz/recidiviz-data#23153, the current implementations of
`gen_cohort_status_df` and `gen_aggregated_cohort_event_df` are such
that the full cohort is not necessarily being used for every time index
in the calculation. In addition to adding a warning when the full cohort
isn't used, this PR also adds a flag to drop any instances that don't
have full observability for the entire time index, so that the same
cohort is used at all time slices (per Recidiviz/recidiviz-data#32590).

This also adds tests for `cohort_methods` for the first time, though the
tests only cover the functionality in this PR for now.

I've shamelessly stolen from Ricardo Zamora's Recidiviz/recidiviz-data#21965, which did this and
more but was never merged. This PR only does a subset of Recidiviz/recidiviz-data#21965
(addressing full observability and adding tests).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#23153
Closes Recidiviz/recidiviz-data#32590
Advances Recidiviz/recidiviz-data#21934

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 04b3f29148ce506e49233255351e034764dfefbc