[Justice Counts] Migrate Single-Page Upload Format Logic to WorkbookUploader (Recidiviz/recidiviz-data#32620)
## Description of the change

I migrated the logic that transforms a single-page format upload from
WorkbookUploader to WorkbookStandardizer. The transformation logic is
the same: we convert the single-page format into a DataFrame that
represents the standard format, which is an Excel workbook with multiple
sheets, each corresponding to a specific metric or disaggregation.
Additionally, I added specific error handling for the single-page
upload. If the single-page format is missing entire columns, an error
will be raised, and the transformation process will not proceed,
ensuring that incomplete data does not move forward in the workflow.
These changes are only for single-page uploads. I will follow-up with a
PR that checks the columns for multi-page uploads early next week.

https://github.com/user-attachments/assets/2e7626b8-b8af-49cf-9aef-4e01f6445a1f

## Related issues

Closes Recidiviz/recidiviz-data#32249

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3dbdc9495e112e3372809cb254bde736a374af59