share logic between supervision_staff_record and current_staff (Recidiviz/recidiviz-data#32652)
## Description of the change

The workflows staff record was recently updated to read from
`supervision_staff_attribute_sessions` and apply some district
formatting logic to match the admin panel; this moves that logic into
the `current_staff` view that `ingested_product_users` also pulls from
and reads from that table in the workflows `supervision_staff_record` so
we don't need to duplicate it.

We need this logic in both views so that districts line up between the
admin panel roster and the workflows staff record, because users without
a staff record need to be able to search for staff in the exact district
specified in the admin panel. Eventually we'll be able to remove a lot
of the formatting logic, but for now it's easier to review diffs if we
keep things the same as they have been.

(also I totally want to look into using this view for insights in the
future! maybe after the district transformation logic has been removed)

Loaded into sandbox at `dana_current_staff`. No diffs to
`ingested_product_users`, and the only diffs to
`supervision_staff_record` are to remove some users who had null
attributes across the board.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/32585

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 92609a7e1fa3de5f1f808619bda01503f5453134