[CI] Move emulator job timeout to step-level (Recidiviz/recidiviz-data#32732)
## Description of the change

I've been seeing a lot of emulator CI jobs fail on PRs lately due to
timeouts that appear to happen when the python setup step takes longer
than expected, leaving less time for the actual tests to run. This
change just moves the timeout to the actual emulator test steps so that
the timeout is less flaky.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Example run with errors here:
https://github.com/Recidiviz/recidiviz-data/actions/runs/10561960212/job/29258853744?pr=32712

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 76f67e242d98b322c4fb0625282e413753baffe8