[Case Note Search] Require state_code and user_id as endpoint params. (Recidiviz/recidiviz-data#32717)
## Description of the change

Require state_code and user_id as endpoint params.

* user_id will be used by the custom logging table (not used yet since
the custom logging table code is waiting for the table migrations before
we can merge it) PR Recidiviz/recidiviz-data#32343

* state_code is used to filter case notes by state, since it's looking
like we will store all case notes in the same datastore, regardless of
state.

## Testing
Tested this locally for the BigQuery data.
Cannot test this on the fake data yet since fake data development is
blocked...

GitOrigin-RevId: 474e4701b899e3afc87b5e360d0ce5accac766c8