[US_ND] Remove usnd_work_release from validation view (Recidiviz/recidiviz-data#32618)
## Description of the change

We changed the name of the opportunity from work-release to ATP. Since
then, we've been failing [this validation
](https://admin-panel-staging.recidiviz.org/admin/validation_metadata/status/details/client_and_resident_record_percent_change_in_eligibility_exceeded?stateCode=US_ND)
for months. This change removes opportunities from the validation if
they are no longer in use (i.e. in
`reference_views.workflows_opportunity_configs`).

The only rows we're dropping are those where `region_code = 'sandbox'`
or where `opportunity_type = 'usNdWorkRelease'` (the one we were hoping
to drop).

![Screenshot 2024-08-26 at 5 38
37 PM](https://github.com/user-attachments/assets/4094e67e-6571-40f5-90b9-454ec554004f)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f42a569ad46098f6880712d8ffcd2fc736cba275