[raw data import] add get_metadata_for_all_raw_files_in_region to metadata manager (Recidiviz/recidiviz-data#32695)
## Description of the change

adds `get_metadata_for_all_raw_files_in_region` to
`DirectIngestRawFileMetadataManagerV2`.

the idea behind this is that since the metadata we are displaying on the
[ingest_raw_data](https://admin-panel-staging.recidiviz.org/admin/ingest_operations/ingest_pipeline_summary/US_ND/instance/PRIMARY#ingest_raw_data)
page is similar to the metadata that we will be displaying for the raw
data import dag tables

i added one additional attribute,
`DirectIngestRawFileMetadataSummary.num_ungrouped_files`, which defaults
to 0 for all non-raw data-import-dag-based-objects. the idea here is
that we'll return the same object for both v1 and v2 managers and only
have a single `ingest_raw_data` page that will display the same
information

once Recidiviz/recidiviz-data#32710 is merged i'll add a follow-on pr for adding most recent
import dag run status

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prework for Recidiviz/recidiviz-data#32694

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5be304f70c6ba9b91cc8c448e5952235ea59088d