[US_PA][Workflows] Add 402a form information for admin supervision opportunity (Recidiviz/recidiviz-data#32775)
## Description of the change

This adds information to fill out the checkboxes in [form
402a](https://docs.google.com/document/d/13m8k6gNsdjG2M_8S3EG7xZG2Z_UOhbQj/edit)
(the drug addendum for the admin supervision opportunity)

When form_information_drug_charge is true, the first checkbox in section
1 should be checked. Also, a flag should display prompting the user to
fill out form 402a.
When form_information_statue_14 is true, the first checkbox in section 2
should be checked.
When form_information_statue_30 is true, the second checkbox in section
2 should be checked.
When form_information_statue_37 is true, the third checkbox in section 2
should be checked.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#31456

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1c3c2faf79f96e7e6d1aae2a2392ca7b9b0d0b2d